Fix compatibility with TeXlive 2016 fontspec
Previously the \spacedallcaps and \spacedlowsmallcaps macros used
the \addfontfeature command to change the letterspace. In newer releases
of fontspec, this globally affects the current font.

See also: https://tex.stackexchange.com/questions/290937/new-version-of-fontspec-gives-wordspace-error/290947

To work around this issue we simply define a new font family that has
all the feature we want enabled.

Please note that this commit uses the calling convention

  \newfontfamily\fontname[<options>{font}

instead of the newer

  \newfontfamily\fontname{font}[<options>]

to ensure compatibility with fontspec < 2.4.